app/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.34
databricks-sdk>=0.36
pydantic>=2.0
databricks-zerobus-ingest-sdk>=0.2.0
azure-eventhub>=5.11.0
asyncpg>=0.29.0
requests>=2.31.0
